Fix log of processKext
Don't log unless the kext is processed.
Indicate in the log which kext is being processed.
return true when a kext is processed.
Call processKext of all patch classes even though each kext is probably only handled by one patch class. This is a minor change because all patch classes are called anyway if none of them handle the kext.